+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add [TextOverflowWidget.position] to support show overflow at start, middle or end.
 * Add [ExtendedText.betterLineBreakingAndOverflowStyle] to make line breaking and overflow style better.
 * Fix strutStyle not work.
+* Breaking change: remove [TextOverflowWidget.fixedOffset]
 ## 5.0.5
 
 * fix strutStyle miss